+++ src/simpandas/__init__.py
@@ -5,9 +5,9 @@ Created on Sun Oct 11 11:14:32 2020
 @author: Martín Carlos Araya <martinaraya@gmail.com>
 """
 
-__version__ = '0.90.0'
-__release__ = 20260420
-__all__ = ['SimSeries', 'SimDataFrame', 'read_excel', 'read_csv', 'read_json',
+__version__ = '0.90.5'
+__release__ = 20260421
+__all__ = ['SimSeries', 'SimDataFrame', 'ColumnUnits', 'read_excel', 'read_csv', 'read_json',
            'read_hdf5', 'read_summary', 'read_vdb', 'read_parquet',
            'read_prodml', 'read_witsml', 'read_resqml', 'concat']
 
@@ -17,3 +17,4 @@ from .index import SimIndex
 from .readers import (read_excel, read_csv, read_json, read_hdf5, read_summary,
                       read_vdb, read_parquet, read_prodml, read_witsml, read_resqml)
 from .common.merger import concat
+from .common.units import ColumnUnits